Fix argument escaping for scp on Windows
