+++ commands/scp.go
@@ -7,7 +7,6 @@ import (
 	"os/exec"
 	"strings"
 
-	"github.com/docker/machine/libmachine"
 	"github.com/docker/machine/libmachine/log"
 	"github.com/docker/machine/libmachine/persist"
 )
@@ -52,26 +51,6 @@ func (s *storeHostInfoLoader) load(name string) (HostInfo, error) {
 	}
 
 	return host.Driver, nil
-}
-
-func cmdScp(c CommandLine, api libmachine.API) error {
-	args := c.Args()
-	if len(args) != 2 {
-		c.ShowHelp()
-		return errWrongNumberArguments
-	}
-
-	src := args[0]
-	dest := args[1]
-
-	hostInfoLoader := &storeHostInfoLoader{api}
-
-	cmd, err := getScpCmd(src, dest, c.Bool("recursive"), c.Bool("delta"), c.Bool("quiet"), hostInfoLoader)
-	if err != nil {
-		return err
-	}
-
-	return runCmdWithStdIo(*cmd)
 }
 
 func getScpCmd(src, dest string, recursive bool, delta bool, quiet bool, hostInfoLoader HostInfoLoader) (*exec.Cmd, error) {
